:recycle: Extract SettingMenstruationPage
